Alle bekannten Bugs die zum Absturz der App geführt haben wurden gefixt. Login/Signup bugs ebenfalls gefixt.

+++ app/src/main/java/com/example/menti/util/EventAdapter.kt
@@ -1,8 +1,11 @@
 package com.example.menti.util
 
+import android.app.AlertDialog
+import android.content.Context
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.example.menti.FirebaseViewModel
 import com.example.menti.data.model.Event
@@ -12,6 +15,7 @@ import com.example.menti.databinding.EventListItemBinding
 class EventAdapter(
     var dataset: MutableList<Event>,
     var firebaseViewModel: FirebaseViewModel,
+    var context: Context
 
 
     ) : RecyclerView.Adapter<EventAdapter.ItemViewHolder>() {
@@ -36,18 +40,29 @@ class EventAdapter(
 
         holder.binding.stornierenBTN.setOnClickListener {
 
-            try {
+            val builder = AlertDialog.Builder(context)
+            builder.setTitle("Achtung")
+            builder.setMessage("Möchtest du diesen Termin wirklich stornieren?")
+            builder.setPositiveButton("Stornieren") { dialog, which ->
+                try {
 
-                firebaseViewModel.deleteEvent(item.id)
-                if (position != RecyclerView.NO_POSITION) {
-                    dataset.removeAt(position)
-                    notifyItemRemoved(position)
+                    firebaseViewModel.deleteEvent(item.id)
+                    if (position != RecyclerView.NO_POSITION) {
+                        dataset.removeAt(position)
+                        notifyItemRemoved(position)
+                        Toast.makeText(context, "Dein Termin wurde Storniert.", Toast.LENGTH_SHORT).show()
+                    }
+                } catch (e: Exception) {
+
+                    Log.e("EventAdapter", e.message.toString())
+
                 }
-            } catch (e: Exception) {
-
-                Log.e("EventAdapter", e.message.toString())
-
             }
+            builder.setNegativeButton("Abbrechen") { dialog, which ->
+                // Aktion beim Klicken auf die Abbrechen-Schaltfläche
+            }
+            builder.show()
+
         }
 
     }
